# Improve combineApiRequests performance