--- a/src/integrations/terminal/__tests__/TerminalProcessExec.pwsh.test.ts
+++ b/src/integrations/terminal/__tests__/TerminalProcessExec.pwsh.test.ts
@@ -225,8 +225,6 @@ describePlatform("TerminalProcess with PowerShell Command Output", () => {
 	beforeAll(() => {
 		// Initialize TerminalRegistry event handlers
 		TerminalRegistry.initialize()
-		// Log environment info
-		console.log(`Running PowerShell tests with PowerShell Core available: ${hasPwsh}`)
 	})
 
 	beforeEach(() => {