## Context

I was trying to reproduce the "Blank chat window of death" issue by examining bottle necks for a huge chat history.  

combineApiRequests was where it become unresponsive for my 200MB - 600m token test.  

With this update, my unrealistic task history data loads properly, so I can now properly reproduce the Blank chat window of death if I drag the scrollbar too high and virtuoso freaks out. (the good news is that basic scrolling still worked well)

It's not the full fix for the unresponsive issue (Which I believe it contributed to by the bottom setting in Virtuoso for large tasks), but it's an improvement.

## Implementation

I had Roo write up some tests for the original combineApiRequests, then write a new, better performing version of the function. (And reviewed of course)

## Screenshots

Because I used Claude, it went wild and created some simple benchmarks:
For 100 API Requests: 1.45x faster
For 500 API Requests: 2.14x faster
For 1,000 API Requests: 3.44x faster
For 100,000 API Requests: 37x faster (this is essentially what my test task history contained)

## How to Test

Load any task from task history

<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Refactor `combineApiRequests` to improve performance and add comprehensive tests for various cases.
> 
>   - **Performance Improvement**:
>     - Refactor `combineApiRequests` in `combineApiRequests.ts` to improve performance from O(n^2) to O(n).
>     - Uses a single pass with a stack to track `api_req_started` messages and combine them with `api_req_finished` messages.
>   - **Behavior**:
>     - Combines `api_req_started` and `api_req_finished` messages into a single message with merged JSON data.
>     - Preserves non-API messages and handles interleaved API and non-API messages.
>     - Filters out `api_req_finished` messages from the result.
>   - **Testing**:
>     - Adds comprehensive tests in `combineApiRequests.test.ts` to verify basic functionality, edge cases, and bug fixes.
>     - Tests include handling of empty arrays, missing text fields, multiple matches, and additional properties.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=RooVetGit%2FRoo-Code&utm_source=github&utm_medium=referral)<sup> for 14188232c104f7a709570d941c687bb79a525d1b. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->